Rename files to use the singular
